move things around to make things tidier
